Moved MainActivity UI to Compose and cleaned up old XML-dependent code
MainActivity, including the NavDrawer, is now implemented in Compose rather
than XML layout files. Placeholders created for Settings and Share options
within drawer (Share btn logs a message when clicked, and Settings screen
is a very rough Composable to expose the logout option.) Since most of the
EventAdapter and ViewHolder classes for the tracker list is no longer used,
deleted a lot of the contents there but left some logic for the delete functionality
so I can reference it later when adding that back in. BaseActivity is now the
activity that's started on app launch, and will be the entry point into the
app.